move install wizard into admin
