+++ src/sentry/static/sentry/app/views/admin/installWizard/index.tsx
