+++ tests/js/spec/views/admin/installWizard.spec.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import {mount} from 'sentry-test/enzyme';
+import {mountWithTheme} from 'sentry-test/enzyme';
 
 import ConfigStore from 'app/stores/configStore';
-import InstallWizard from 'app/views/installWizard';
+import InstallWizard from 'app/views/admin/installWizard';
 
 jest.mock('app/stores/configStore', () => ({
   get: jest.fn(),
@@ -28,7 +28,7 @@ describe('InstallWizard', function() {
   beforeEach(function() {});
 
   it('renders', function() {
-    const wrapper = mount(<InstallWizard onConfigured={jest.fn()} />);
+    const wrapper = mountWithTheme(<InstallWizard onConfigured={jest.fn()} />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -49,7 +49,7 @@ describe('InstallWizard', function() {
         },
       }),
     });
-    const wrapper = mount(<InstallWizard onConfigured={jest.fn()} />);
+    const wrapper = mountWithTheme(<InstallWizard onConfigured={jest.fn()} />);
 
     expect(
       wrapper.find('input[name="beacon.anonymous"][value="false"]').prop('checked')
@@ -77,7 +77,7 @@ describe('InstallWizard', function() {
         },
       }),
     });
-    const wrapper = mount(<InstallWizard onConfigured={jest.fn()} />);
+    const wrapper = mountWithTheme(<InstallWizard onConfigured={jest.fn()} />);
 
     expect(
       wrapper.find('input[name="beacon.anonymous"][value="false"]').prop('checked')
